requirements.txt
torch>=2.0.0
numpy>=1.20.0
pyyaml>=6.0
tqdm>=4.60.0
